Propagate parameter annotations in generated factory code.
Example:
```java
@AutoFactory
final class SomeClass {
  ...
  SomeClass(@Provided @AQualifier @Foo String provideDepA, @Bar String depB) {...}
}
```

The generated `SomeClassFactory` will have a `create` method like this:
```java
  SomeClass create(@Bar String depB) {...}
```

This was already true if `@Bar` was a `TYPE_USE` annotation, but now it is also true for other annotations.

This may affect compilation. For example, if `@Bar` is Kotlin's `@NotNull`,
this change plugs a hole in Kotlin's nullness checking and will break callers
that were previously passing a `@Nullable` value. The generated factory already
null-checked such parameters, so in Kotlin you can add `!!` with no change in
effect. If the `!!` causes a `NullPointerException`, there would already have
been a `NullPointerException` from the generated factory.

We do not currently copy annotations, other than `@Qualifier` annotations, from `@Provided` parameters to the generated factory constructor. In the example, `@Foo` would not be copied.

PiperOrigin-RevId: 481174930