Use a narrowing cast with `|=` where appropriate.
We use bitmasks to record which required properties have been set. The size of the bitmask variable depends on the number of required properties. For example, if there are 8 or fewer we use a `byte`. Before this change, we set bits with expressions like `set$0 |= 1`. That's legal when `set$0` is a `byte`, even though `set$0 = 1` would not be. It's also clearly safe when the RHS fits in a byte, as it always does in our case. Nevertheless, some static code checkers warn about constructs like this. It's straightforward just to cast and remove any doubt, `set$0 |= (byte) 1`.

Fixes https://github.com/google/auto/issues/1384.

RELNOTES=n/a
PiperOrigin-RevId: 490255211