Loosen the rules for AutoBuilder copy-constructors.
If the built type has a method `T getFoo()` and its constructor has a parameter `U foo`, then we should still generate a copy-constructor provided `T` is assignable to `U`. The copy-constructor will call `getFoo()` on its parameter and by default that is the value that will get passed to the constructor when `build()` is called.

RELNOTES=AutoBuilder copy-constructors no longer require an exact match between a property and the corresponding constructor parameter.
PiperOrigin-RevId: 508917172