Arrange for the Gradle integration test to run in the `install` phase.
I am not sure how this test worked before. I think it happened to work as a side effect of other integration tests installing the `HEAD-SNAPSHOT` artifacts, which we stopped doing in https://github.com/google/auto/pull/1451. It happens that we noticed the breakage with a Dependabot [PR](https://github.com/google/auto/pull/1452) to update `gradle-test-kit`, but I think the CI was already broken.

Anyway, the Gradle test assumes it can depend on the artifacts for the current AutoValue version, and those are only available in the `install` phase.

By following the Maven convention whereby regular unit tests are called `FooTest` and integration tests are called `FooIT`, we get the right tests run in the right phases without having to specify `<include>` and the like.

RELNOTES=n/a
PiperOrigin-RevId: 511051082