Clarify which argument is meant in an exception message.
"Argument index 1" could plausibly mean the second argument, but "Argument 1 of 3" is more obviously the first. It also makes it easier to count backwards from the end if there are a lot of arguments.

Fixes https://github.com/google/auto/issues/1495.

RELNOTES=The exception message for a null value now indicates not only which number argument it is but also how many arguments there are in total. This may lead to a small increase in code size.
PiperOrigin-RevId: 529210616