Add a test to reproduce a problem with annotations and modules.
If your AutoValue class inherits an abstract method from a module, and that abstract method has an annotation that is not exported from the module, then AutoValue will incorrectly try to copy the annotation onto the generated implementation of the abstract method.

In a later change we plan to fix this problem and invert the sense of the test.

RELNOTES=n/a
PiperOrigin-RevId: 532139999