Output a warning if a `setX` method in a `Builder` is marked `@Nullable`.
If the `setX` method itself, or its return type, is `@Nullable` then a warning makes sense. The method always returns a non-null value, the `Builder` itself.

RELNOTES=A warning is now produced if a `setX` method in a `Builder` or its return type is marked `@Nullable`. Those methods always return the `Builder` instance, which is never null.
PiperOrigin-RevId: 547329146