Internal change.
This change adds logic to avoid copying an annotation to the generated subclass if the annotation is not visible due to being in another module and not exported. For now it is excluded from the public version because we still build on Java 8 and it would be a big pain to add the necessary reflective logic. We will probably stop supporting Java 8 in future versions.

RELNOTES=n/a
PiperOrigin-RevId: 565673767