The "copy constructor" in a generated `Builder` is no longer private.
This is the constructor that is generated when the containing `@AutoValue` class has an abstract `toBuilder()` method. The implementation of that method calls the copy constructor. By making the constructor visible, we allow extensions to subclass `Builder` and call its copy constructor.

RELNOTES=The "copy constructor" in a generated `Builder` is no longer private.
PiperOrigin-RevId: 570125710