Cosmetic indentation tweaks to `builder.vm`.
The changes have no effect on the generated code, because it gets reindented correctly anyway after template substitution. But it makes reading the template very slightly easier. For example the assignment to `missing` is inside an `if` statement, so it should be indented further than that `if` statement.

(In these templates, there are two separate kinds of indentation: the lines of generated code have an appropriate code indentation, and the Velocity directives like `#if` and `#foreach` have their own indentation reflecting their nesting.)

RELNOTES=n/a
PiperOrigin-RevId: 581966992