AutoValue support for very large classes.
This change makes it so that if the AutoValue class defines a Builder method, the builder will be passed to the constructor. Effectively working around the java 255 method-arg limit.

This is instead of defining a constructor where every field in the autovalue is an argument in it.

This change will only apply to autovalues that don't have extensions: in which case the constructor is private and we can be assured that updating it shouldn't be problematic for current cases.

RELNOTES=Support for AutoValues with +255 properties.
PiperOrigin-RevId: 582096324