Don't call `Types.asMemberOf` unnecessarily.
If method `m` is defined in class `C`, there's no point in calling `asMemberOf(C, m)`. `asMemberOf` is only useful when `m` is defined in a subtype of `C`.

This is a very minor optimization, but it also avoids an issue where `asMemberOf` can return a method that has lost type annotations on its return type. We still have that issue for inherited methods, but at least in the common case of locally-defined methods we can avoid it.

RELNOTES=Annotations on type parameters, like `abstract @Nullable T foo()`, are now better propagated to fields and constructor parameters.
PiperOrigin-RevId: 587175179