In an AutoValueJava8Test test method, accept any `@Nullable` annotation.
This test has a local `@Nullable` that is nested in the test class itself, and applied to the type of a property method in an AutoValue class being tested. Previously we expected that that annotation would also be copied onto the parameter type of `equals(Object)`. But if `java.lang.Object` already has a `@Nullable` on its `equals` parameter, we'll copy that instead. So the test now accepts either.

RELNOTES=n/a
PiperOrigin-RevId: 587759044