Remove support from AutoValue for generating Java 7 code.
We now only support Java 8. 1.10.4 was the last AutoValue release that could generate Java 7 code.

Fixes #1683.

RELNOTES=Support for generating Java 7 code has been removed from AutoValue, AutoAnnotation, and AutoBuilder. You must be on at least Java 8, or an Android version with desugaring that allows it to pass for Java 8. 1.10.4 is the last AutoValue version with support for Java 7.
PiperOrigin-RevId: 597954833