Copy `Truth8.assertThat` overloads for `Optional` and `Stream` to the main `Truth` class.
(and prepare AutoValue tests for the new overloads, since the Eclipse compiler considers the (temporary!) signatures of `assertThat(Optional)` to be ambiguous)

We'll post some migration suggestions as part of the release notes.

This is the biggest part of https://github.com/google/truth/issues/746.

RELNOTES=Added `assertThat` overloads for `Optional` and `Stream` to the main `Truth` class.
PiperOrigin-RevId: 598664192