In `BasicAnnotationProcessor`, correctly defer non-type elements rejected by processing steps.
See https://github.com/google/auto/pull/1348 for details. If a processor is triggered by a method annotation, for example, and the processor rejects the method until the next round, then we should defer processing of just that method, not the entire class containing it.

Closes https://github.com/google/auto/pull/1348.

RELNOTES=`BasicAnnotationProcessor` has improved handling of deferring processing of methods and other non-type elements. Thanks to @AriaAdibi for this contribution!
PiperOrigin-RevId: 603138183