Fix the logic for computing the number of `int` default-value bitmasks in a Kotlin constructor.
There are ⌈_n_/32⌉ `int` bitmasks when the total number of parameters is _n_. Previously the logic was based on _n_ being the number of _optional_ parameters (the ones with default values).

RELNOTES=A bug with AutoBuilder and Kotlin data classes has been fixed. If there was a mix of required and optional parameters in a data class with a large number of properties, sometimes the generated code would not compile.
PiperOrigin-RevId: 651226198