Ensure type names in AutoOneOf generated code are qualified if needed.
If the parent class has a nested type called `String` or `Object` then that is
what plain `String` or `Object` in the generated code will refer to. That's
pretty unlikely to happen, but we should still handle it properly. The use of
backticks in the template means that the fully-qualified name will be used if
necessary.

RELNOTES=n/a
PiperOrigin-RevId: 660047188