Migrate from `MoreStreams` to equivalent methods on `ImmutableList` and friends.
The equivalent methods on `ImmutableList` and friends are now safe to use, even when a target wrongly brings the Android flavor of Guava onto the annotation-processor classpath. Those methods are present in the Android flavor of Guava since [33.2.0](https://github.com/google/guava/releases/tag/v33.2.0), so this CL is roughly as safe as another CL that begins using Guava methods that were introduced 4 months ago.

RELNOTES=n/a
PiperOrigin-RevId: 672554534