Fix an AutoValue crash with `@AutoValue.CopyAnnotations(excluded=Missing.class)`.
If `Missing.class` is a class that does not exist, then with javac, at least, the `AnnotationValue` for it does not contain an `ErrorType`, as you might expect, but a `String`. We now ignore anything that is not a `DeclaredType`. Presumably the missing class will result in a compilation error anyway.

Fixes https://github.com/google/auto/issues/1901.

RELNOTES=Fixed a crash with `@AutoValue.CopyAnnotations(exclude=Missing.class)`, where `Missing` is a class that does not exist.
PiperOrigin-RevId: 739196666