Add a little clarification about static factory methods versus record constructors.
The text here was intended to be about the common case where your AutoValue class `Foo` has a single static factory method which calls the generated `AutoValue_Foo` constructor. Then if you replace `Foo` with a record, there's not a whole lot of benefit to maintaining the method, apart from easing migration. But one benefit is that methods have names. This is mostly relevant when you have _several_ static factory methods.

RELNOTES=n/a
PiperOrigin-RevId: 740873396