Remove code from `AutoServiceProcessor` that read existing `META-INF/services` files from the `CLASS_OUTPUT` location and merged new `@AutoService` services with them.
This behavior was untested and it's unclear what situations would actually need this behavior. Internally, nothing was broken by removing this and continuing to include it was causing problems for reasons I won't get into. If it turns out that open-source users _were_ relying on this behavior we can add it back in with some way of turning it off.

RELNOTES=Removed `AutoServiceProcessor` code that read existing `META-INF/services` files.
PiperOrigin-RevId: 743977726